LD-19: Add functionality for deleting material submissions

+++ app/Services/MaterialFile/MaterialFileSubmissionService.php
@@ -5,9 +5,11 @@ namespace App\Services\MaterialFile;
 use App\Models\Enums\MaterialSubmissionStatus;
 use App\Models\Enums\SubmissionType;
 use App\Models\MaterialFile;
+use App\Models\MaterialFileState;
 use App\Models\MaterialFileSubmission;
 use App\Models\MaterialVersionSubmission;
 use App\Services\MaterialFile\Dto\MaterialFileSubmissionDto;
+use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Facades\Auth;
 
 readonly class MaterialFileSubmissionService
@@ -58,21 +60,83 @@ readonly class MaterialFileSubmissionService
 
     public function delete(MaterialFileSubmission $fileSubmission): void
     {
-        $isUnreleasedFile = $fileSubmission->type === SubmissionType::Create
-            && $fileSubmission->versionSubmission->materialSubmission->status !== MaterialSubmissionStatus::Accepted;
-        $isRemovedFile    = $fileSubmission->type === SubmissionType::Delete
-            && $fileSubmission->versionSubmission->materialSubmission->status === MaterialSubmissionStatus::Accepted;
-
-        if ($isUnreleasedFile || $isRemovedFile) {
+        if ($this->shouldDeleteFile($fileSubmission)) {
             $this->fileService->delete($fileSubmission->file);
             return;
         }
 
-        if ($fileSubmission->type === SubmissionType::Delete) {
+        if ($this->canDeleteState($fileSubmission)) {
+            $this->fileStateService->delete($fileSubmission->fileState);
+        } else {
             $fileSubmission->delete();
-            return;
         }
 
-        $this->fileStateService->delete($fileSubmission->fileState);
+        $this->deletePreviousStateIfNeeded($fileSubmission);
+    }
+
+    private function shouldDeleteFile(MaterialFileSubmission $fileSubmission): bool
+    {
+        $isUnreleased = $fileSubmission->type === SubmissionType::Create
+            && $fileSubmission->status !== MaterialSubmissionStatus::Accepted;
+        $isRemoved    = $fileSubmission->type === SubmissionType::Delete
+            && $fileSubmission->status === MaterialSubmissionStatus::Accepted;
+
+        return $isUnreleased || $isRemoved;
+    }
+
+    private function canDeleteState(MaterialFileSubmission $fileSubmission): bool
+    {
+        if ($fileSubmission->type === SubmissionType::Delete) {
+            return false;
+        }
+
+        if ($fileSubmission->status === MaterialSubmissionStatus::Accepted) {
+            $latestFileStateId = MaterialFileState::where('file_id', $fileSubmission->file_id)
+                ->orderByDesc('published_at')
+                ->value('id');
+
+            if ($latestFileStateId === $fileSubmission->file_state_id) {
+                return false;
+            }
+
+            $isStatePreviousForAnySubmission = MaterialFileSubmission
+                ::join('material_version_submissions', 'material_version_submissions.id', '=', 'material_file_submissions.version_submission_id')
+                ->join('material_submissions', 'material_submissions.id', '=', 'material_version_submissions.material_submission_id')
+                ->where('file_id', $fileSubmission->file_id)
+                ->whereIn('material_submissions.status', [MaterialSubmissionStatus::Accepted, MaterialSubmissionStatus::Rejected])
+                ->where('material_submissions.updated_at', '>', $fileSubmission->fileState->published_at)
+                ->where(function (Builder $query) use ($fileSubmission) {
+                    // id последнего опубликованного состояния на момент до закрытия заявки
+                    $query->select('id')
+                        ->from('material_file_states')
+                        ->whereNotNull('published_at')
+                        ->whereColumn('material_file_states.file_id', 'material_file_submissions.file_id')
+                        ->whereColumn('material_file_states.published_at', '<', 'material_submissions.updated_at')
+                        ->orderByDesc('material_file_states.published_at')
+                        ->limit(1);
+                }, $fileSubmission->file_state_id)
+                ->exists();
+
+            if ($isStatePreviousForAnySubmission) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function deletePreviousStateIfNeeded(MaterialFileSubmission $fileSubmission): void
+    {
+        if ($fileSubmission->is_closed) {
+            $previousState = MaterialFileState::where('file_id', $fileSubmission->file_id)
+                ->where('published_at', '<', $fileSubmission->updated_at)
+                ->orderByDesc('published_at')
+                ->withExists('fileSubmission')
+                ->first();
+
+            if ($previousState !== null && !$previousState->file_submission_exists) {
+                $this->fileStateService->delete($previousState);
+            }
+        }
     }
 }
